feat(core): centralize plugin loading via PluginManager
