+++ .changeset/centralize-plugin-management.md
@@ -1,0 +1,5 @@
+---
+'@lapidist/design-lint': patch
+---
+
+refactor plugin loading with centralized PluginManager
